refactor: extract SqliteBaseRepository to deduplicate engine/session init
The three SQLite repositories (task, notes, audit_log) duplicated
identical __init__/close() logic (~25 lines each). Extract shared
engine management and session factory initialization into
SqliteBaseRepository base class.

Also extract _create_builders() helper in SqliteTaskRepository to
reduce duplicated builder initialization between create() and
save_all().
